requirements.txt
streamlit>=1.29
pydantic>=2.5
pandas>=2.1
requests>=2.31
python-dotenv>=1.0
code-index-mcp>=0.3
GitPython>=3.1
PyGithub>=2.3
beautifulsoup4>=4.12
lxml>=4.9
python-docx>=0.8
pdfplumber>=0.10
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10 